fix(augmentation): auto-augment ops preserve float16 / bfloat16 input dtype (#4210)
* fix(augmentation): auto-augment ops preserve float16 / bfloat16 input dtype

OperationBase.forward moved the batch_prob gate to input.device but not
input.dtype, so batch_prob (float32) * op(x) + (1 - batch_prob) * x promoted a
float16 / bfloat16 input to float32. float32 / float64 were unaffected because
promotion goes up. This gate is shared by RandAugment / AutoAugment /
TrivialAugment / AugMix, so all of them silently upcast low-precision inputs.

Cast batch_prob to input.dtype as well as input.device. float32 output is
byte-identical. Adds test_operation_preserves_input_dtype (fails on main for
float16 and bfloat16, passes with this change).

* Drop now-passing half-precision xfails; skip torch-version dtype gaps

The fix makes 9 float16 + 10 bfloat16 auto-augment entries in
testing/half_precision_xfails/{cpu_float16,cpu_bfloat16}.txt pass (they were
RuntimeError xfails tracked in #4153), so CI's strict-xpass check flagged them --
remove those entries.

test_operation_preserves_input_dtype now skips an op that raises
'... not implemented for Half/BFloat16' (grid_sample has no half CPU kernel on
older torch, e.g. 2.5.1) -- a torch capability gap, not a kornia contract. Still
fails on unfixed main (AutoContrast: float32 != float16), still passes with the fix.

* half-precision xfails: keep test_reproduce[cpu-float16], failure mode RuntimeError -> AssertionError

With the batch_prob dtype fix, auto-augment runs in fp16 instead of crashing, but
TestAutoAugment::test_reproduce still fails on the half-precision CI (torch 2.9.1):
re-running with recorded params differs above the fp16 rtol/atol of 1e-3 (an
existing fp16 exactness fragility this fix merely unmasks; it passes on torch
2.11). Re-add it as a builtins.AssertionError xfail. The other 8 float16 + all 10
bfloat16 auto_operation xfails genuinely pass and stay removed.

* test: gate dtype-preservation check to CPU

A freshly built _find_all_ops() op's auto-augment RNG state does not fully
follow a post-hoc .to(mps), causing a cpu/mps device-mismatch RuntimeError on
the mps CI job. dtype preservation is device-independent, so restrict the check
to CPU (also drops the now-unnecessary op.to(device) / x device kwarg). Still
fails on unfixed main, still passes with the fix.

* CHANGELOG: auto-augment half-precision dtype fix (#4210)

+++ kornia/augmentation/auto/operations/base.py
@@ -192,7 +192,9 @@ class OperationBase(nn.Module):
         if params is None:
             params = self.forward_parameters(input.shape)
 
-        batch_prob = params["batch_prob"][(...,) + ((None,) * (len(input.shape) - 1))].to(device=input.device)
+        batch_prob = params["batch_prob"][(...,) + ((None,) * (len(input.shape) - 1))].to(
+            device=input.device, dtype=input.dtype
+        )
 
         return batch_prob * self.op(input, params=params) + (1 - batch_prob) * input
